Turbine Generator now uses a double to track its velocity Tweaked energy generation values of Furnace Generator Fixed tx.commit() not being called where it should have been... Sonion ring.

+++ src/main/java/net/myriantics/klaxon/mechanics/turbine_generator/boost/TurbineGeneratorBoostInstance.java
@@ -18,14 +18,14 @@ public abstract sealed class TurbineGeneratorBoostInstance permits TurbineGenera
 
     public static final class AddToBase extends TurbineGeneratorBoostInstance {
 
-        private final long addedValue;
+        private final double addedValue;
 
-        public AddToBase(int duration, long addedValue) {
+        public AddToBase(int duration, double addedValue) {
             super(duration);
             this.addedValue = addedValue;
         }
 
-        public long getAddedValue() {
+        public double getAddedValue() {
             return this.addedValue;
         }
     }
